requirements.txt
torch
torchvision
flwr
opacus
gradio
numpy
matplotlib
scikit-learn
pillow
seaborn